Extracted helper utils to separate files
