+++ src/app/library/media/[id]/page.tsx
@@ -4,7 +4,8 @@ import { API_LINKS } from 'app/links';
 import { BASE_URL } from 'app/shared/constants';
 import { Metadata, ResolvingMetadata } from 'next';
 import { T_RawMediaContentFields } from 'types/media-content';
-import { determineFileType, getMetadataGenerator } from 'utils/helpers';
+import { determineFileType } from 'utils/helpers/determineFileType';
+import { getMetadataGenerator } from 'utils/helpers/getMetadataGenerator';
 import NETWORK_UTILS from 'utils/network';
 import { MediaContentView } from '../../../../components/library/MediaContentView';
 import { fetchRelatedMedia } from '../../../../fetchers/library/fetchRelatedMedia';
