+++ src/app/library/search/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import EmptyContentIndicator from '@components/library/EmptyContentIndicator';
-import { getMetadataGenerator } from 'utils/helpers';
+import { getMetadataGenerator } from 'utils/helpers/getMetadataGenerator';
 import { LibraryErrorMessage } from '@components/library/LibraryErrorMessage/LibraryErrorMessage';
 import { T_LibrarySearchPageProps } from '@components/library/types';
 import { Metadata, ResolvingMetadata } from 'next';
