+++ src/components/courses/editCourseView.tsx
@@ -8,7 +8,7 @@ import { FormEventHandler, useState } from 'react';
 import { COURSE_FORM_FIELDS } from './constants';
 import useCourse from '@hooks/useCourse';
 import { AdminFormInput } from '@components/admin/AdminForm/AdminFormInput';
-import { extractValuesFromFormEvent } from 'utils/helpers';
+import { extractValuesFromFormEvent } from 'utils/helpers/extractValuesFromFormEvent';
 import { T_CourseFields } from '@hooks/useCourse/types';
 import { LibraryErrorMessage } from '@components/library/LibraryErrorMessage/LibraryErrorMessage';
 import { DeletionWarningModal } from '@components/admin/AdminTable/DeletionWarningModal';
