+++ src/components/grades/createGradeView.tsx
@@ -9,7 +9,7 @@ import { GRADE_FORM_FIELDS } from './constants';
 import { transformRawProgram } from '@hooks/useProgram';
 import { AdminFormInput } from '@components/admin/AdminForm/AdminFormInput';
 import { AdminFormSelector } from '@components/admin/AdminForm/AdminFormSelector';
-import { extractValuesFromFormEvent } from 'utils/helpers';
+import { extractValuesFromFormEvent } from 'utils/helpers/extractValuesFromFormEvent';
 import { T_CreateGradeFields } from '@hooks/useGrade/types';
 import { useAdminListViewData } from '@hooks/useAdmin/useAdminListViewData';
 import { API_LINKS } from 'app/links';
