+++ src/utils/helpers/capitalizeFirstLetter.ts
@@ -1,0 +1,3 @@
+export function capitalizeFirstLetter(str: string): string {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
