+++ src/utils/helpers/determineFileType.ts
@@ -1,0 +1,24 @@
+export type AcceptableFileType = 'image' | 'video' | 'pdf';
+
+export function determineFileType(fileUrl: string): AcceptableFileType | null {
+  let extension = fileUrl.split('.').pop()?.toLowerCase();
+
+  switch (extension) {
+    case 'jpg':
+    case 'jpeg':
+    case 'png':
+    case 'gif':
+    case 'bmp':
+    case 'webp':
+    case 'svg':
+      return 'image';
+    case 'mp4':
+    case 'webm':
+    case 'ogg':
+      return 'video';
+    case 'pdf':
+      return 'pdf';
+    default:
+      return null;
+  }
+}
