+++ src/utils/helpers/toTitleCase.ts
@@ -1,0 +1,9 @@
+export const toTitleCase = (text: string) => {
+  if (!text) return '';
+
+  return text
+    .toLowerCase()
+    .split(' ')
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+    .join(' ');
+};
